replace all `rsm` with `cxxst` & rename project in `CMakeLists.txt`
