+++ README.md
@@ -1,8 +1,8 @@
-# Runtime simple markers
+# C++ simple tracing
 
 TODO ...
 
-Examples are [here](examples/README.md).
+[Examples are here](examples/README.md).
 
 ## Explorations
 
@@ -32,7 +32,7 @@ Maybe create github issues for those?
 * better & wider test coverage
   * [recommendation](https://github.com/Ruzovej/compare_unit_test_frameworks/blob/main/docs/g++_release.md#doctest-with-macro-as-above--using-binary-asserts) for unit test usage
 * benchmarking
-  * try reimplementing the `rsm::marker` to don't hold all the data, but "save" them immediately and update end time on `submit`/d-ror (by holding only pointer to this particular record)
+  * try reimplementing the `cxxst::marker` to don't hold all the data, but "save" them immediately and update end time on `submit`/d-ror (by holding only pointer to this particular record)
 * add more output formats (e.g. implied by this conversation: <https://chatgpt.com/share/681b2833-c728-8003-9e28-a4e1c1c35196>)
   * read e.g. <https://www.chromium.org/developers/how-tos/trace-event-profiling-tool/> and fix how it's saved -> e.g. ~~"stacked" markers are properly displayed~~
 * [TODO.md](doc/TODO.md)
