+++ easy.bash
@@ -15,9 +15,9 @@ time ./tests.bash --bats
 set -x
 
 export CXXST_DEFAULT_BLOCK_SIZE=3 # torture it with some non-default value & force it to allocate more during the run
-time bin/release/rsm_example_complete_1 /tmp/easy_complete_1.json
-time bin/release/rsm_example_counter_1 /tmp/easy_counter_1.json
-time bin/release/rsm_example_counter_2 /tmp/easy_counter_2.json
-time bin/release/rsm_example_duration_1 /tmp/easy_duration_1.json
-time bin/release/rsm_example_instant_1 /tmp/easy_instant_1.json
-time bin/release/rsm_example_instant_2 /tmp/easy_instant_2.json
+time bin/release/cxxst_example_complete_1 /tmp/easy_complete_1.json
+time bin/release/cxxst_example_counter_1 /tmp/easy_counter_1.json
+time bin/release/cxxst_example_counter_2 /tmp/easy_counter_2.json
+time bin/release/cxxst_example_duration_1 /tmp/easy_duration_1.json
+time bin/release/cxxst_example_instant_1 /tmp/easy_instant_1.json
+time bin/release/cxxst_example_instant_2 /tmp/easy_instant_2.json
