+++ examples/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_custom_target(
-    rsm_examples
+    cxxst_examples
 )
 
 function(add_example target_name source_file)
@@ -11,7 +11,7 @@ function(add_example target_name source_file)
     target_link_libraries(
         ${target_name}
             PRIVATE
-            rsm
+            cxxst
     )
 
     add_executable(
@@ -22,42 +22,42 @@ function(add_example target_name source_file)
     target_link_libraries(
         ${target_name}_bare
             PRIVATE
-            rsm_bare
+            cxxst_bare
     )
 
     add_dependencies(
-        rsm_examples
+        cxxst_examples
             ${target_name}
             ${target_name}_bare
     )
 endfunction()
 
 add_example(
-    rsm_example_instant_1
+    cxxst_example_instant_1
         "instant/1.cpp"
 )
 
 add_example(
-    rsm_example_instant_2
+    cxxst_example_instant_2
         "instant/2.cpp"
 )
 
 add_example(
-    rsm_example_counter_1
+    cxxst_example_counter_1
         "counter/1.cpp"
 )
 
 add_example(
-    rsm_example_counter_2
+    cxxst_example_counter_2
         "counter/2.cpp"
 )
 
 add_example(
-    rsm_example_duration_1
+    cxxst_example_duration_1
         "duration/1.cpp"
 )
 
 add_example(
-    rsm_example_complete_1
+    cxxst_example_complete_1
         "complete/1.cpp"
 )
