+++ examples/README.md
@@ -1,6 +1,6 @@
 # Examples
 
-This directory contains examples demonstrating the usage of the `rsm` library for tracing `C++` code.
+This directory contains examples demonstrating the usage of the `cxxst` library for tracing `C++` code.
 
 It's lightweight library that generates `chrome trace` format output, which can be viewed in Chrome's built-in trace viewer (`chrome://tracing`) or tools like [Perfetto UI](https://ui.perfetto.dev/).
