+++ examples/complete/1.cpp
@@ -1,7 +1,7 @@
 #include <chrono>
 #include <thread>
 
-#include "rsm/all.hpp"
+#include "cxxst/all.hpp"
 
 static void pyramid(int const level) {
   CXXST_mark_complete(__FUNCTION__);
@@ -18,7 +18,7 @@ int main(int argc, char const **argv) {
 
   [[maybe_unused]] char const *const filename{argc > 1 ? argv[1]
                                                        : "/dev/stdout"};
-  CXXST_flush_global_sink(rsm::output::format::chrome_trace, filename, true);
+  CXXST_flush_global_sink(cxxst::output::format::chrome_trace, filename, true);
 
   std::thread t1{[]() {
     CXXST_init_thread_local_sink();
