+++ examples/duration/1.cpp
@@ -1,12 +1,12 @@
 #include <chrono>
 #include <thread>
 
-#include "rsm/all.hpp"
+#include "cxxst/all.hpp"
 
 int main(int argc, char const **argv) {
   [[maybe_unused]] char const *const filename{argc > 1 ? argv[1]
                                                        : "/dev/stdout"};
-  CXXST_flush_global_sink(rsm::output::format::chrome_trace, filename, true);
+  CXXST_flush_global_sink(cxxst::output::format::chrome_trace, filename, true);
 
   CXXST_init_thread_local_sink();
