+++ include/private/impl/dump_records.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
 #include "impl/event/list/list.hpp"
-#include "rsm/output_format.hpp"
+#include "cxxst/output_format.hpp"
 
-namespace rsm::impl {
+namespace cxxst::impl {
 
 void dump_records(impl::event::list const &list,
                   long long const time_point_zero, output::format const fmt,
