+++ include/private/impl/event/kind/complete.hpp
@@ -2,7 +2,7 @@
 
 #include "impl/event/common.hpp"
 
-namespace rsm::impl::event {
+namespace cxxst::impl::event {
 
 // https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/edit?tab=t.0#heading=h.lpfof2aylapb
 
@@ -32,4 +32,4 @@ struct complete {
   }
 };
 
-} // namespace rsm::impl::event
+} // namespace cxxst::impl::event
