+++ include/private/impl/thread_local_sink_submit_event.hpp
@@ -2,7 +2,7 @@
 
 #include "impl/event/any.hpp"
 
-namespace rsm::impl {
+namespace cxxst::impl {
 
 void thread_local_sink_submit_event(impl::event::any const &evt) noexcept;
