+++ include/public/cxxst/macros/implicit_name.h
