+++ include/public/cxxst/macros/linkage.h
