+++ include/public/cxxst/mark_duration_begin.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "rsm/macros/linkage.h"
-#include "rsm/timepoint.hpp"
+#include "cxxst/macros/linkage.h"
+#include "cxxst/timepoint.hpp"
 
-namespace rsm {
+namespace cxxst {
 
 CXXST_IMPL_API void submit_duration_begin(char const *const desc,
                                         long long const timestamp_ns) noexcept;
@@ -12,4 +12,4 @@ inline void mark_duration_begin(char const *const desc) noexcept {
   submit_duration_begin(desc, impl::as_int_ns(impl::now()));
 }
 
-} // namespace rsm
+} // namespace cxxst
