+++ include/public/cxxst/mark_instant.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "rsm/macros/linkage.h"
-#include "rsm/scope.hpp"
-#include "rsm/timepoint.hpp"
+#include "cxxst/macros/linkage.h"
+#include "cxxst/scope.hpp"
+#include "cxxst/timepoint.hpp"
 
-namespace rsm {
+namespace cxxst {
 
 CXXST_IMPL_API void submit_instant(char const *const desc, scope_t const scope,
                                  long long const timestamp_ns) noexcept;
@@ -14,4 +14,4 @@ inline void mark_instant(char const *const desc,
   submit_instant(desc, scope, impl::as_int_ns(impl::now()));
 }
 
-} // namespace rsm
+} // namespace cxxst
