+++ include/public/cxxst/output_format.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-namespace rsm::output {
+namespace cxxst::output {
 
 enum class format : int {
   chrome_trace, // https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview?tab=t.0
   raw_naive_v0,
 };
 
-} // namespace rsm::output
+} // namespace cxxst::output
