+++ src/rsm/mark_counter.cpp
@@ -1,9 +1,9 @@
-#include "rsm/mark_counter.hpp"
+#include "cxxst/mark_counter.hpp"
 
 #include "impl/event/kind/complete.hpp"
 #include "impl/thread_local_sink_submit_event.hpp"
 
-namespace rsm {
+namespace cxxst {
 
 void submit_counter(char const *const name, long long const timestamp_ns,
                     double const value) noexcept {
@@ -11,4 +11,4 @@ void submit_counter(char const *const name, long long const timestamp_ns,
       impl::event::counter{name, timestamp_ns, value});
 }
 
-} // namespace rsm
+} // namespace cxxst
