+++ src/rsm/mark_duration_begin.cpp
@@ -1,9 +1,9 @@
-#include "rsm/mark_duration_begin.hpp"
+#include "cxxst/mark_duration_begin.hpp"
 
 #include "impl/event/kind/duration.hpp"
 #include "impl/thread_local_sink_submit_event.hpp"
 
-namespace rsm {
+namespace cxxst {
 
 CXXST_IMPL_API void submit_duration_begin(char const *const desc,
                                         long long const timestamp_ns) noexcept {
@@ -11,4 +11,4 @@ CXXST_IMPL_API void submit_duration_begin(char const *const desc,
       impl::event::duration_begin{desc, timestamp_ns});
 }
 
-} // namespace rsm
+} // namespace cxxst
