+++ src/rsm/mark_instant.cpp
@@ -1,9 +1,9 @@
-#include "rsm/mark_instant.hpp"
+#include "cxxst/mark_instant.hpp"
 
 #include "impl/event/kind/instant.hpp"
 #include "impl/thread_local_sink_submit_event.hpp"
 
-namespace rsm {
+namespace cxxst {
 
 void submit_instant(char const *const desc, scope_t const scope,
                     long long const timestamp_ns) noexcept {
@@ -11,4 +11,4 @@ void submit_instant(char const *const desc, scope_t const scope,
       impl::event::instant{desc, scope, timestamp_ns});
 }
 
-} // namespace rsm
+} // namespace cxxst
