+++ tests/unit/unit_test_runner.bash
@@ -36,10 +36,10 @@ function unit_test_runner() {
                 ./compile.bash \
                     -DCXXST_BUILD_TESTS=ON \
                     --preset "${preset}" \
-                    --target rsm_unit_tests \
+                    --target cxxst_unit_tests \
                     --polite-ln-compile_commands
             fi
-            "bin/${preset}/rsm_unit_tests" \
+            "bin/${preset}/cxxst_unit_tests" \
                 --no-intro
             printf '\n'
         done
